Data Fetch/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
python-dateutil
isodate
google-api-python-client
google-cloud-storage